requirements.txt
Flask==2.3.3
anthropic==0.61.0
httpx>=0.25.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyMuPDF==1.22.5
gunicorn==21.2.0
lxml==4.9.3
azure-cognitiveservices-speech==1.40.0